ungoogled-chromium: import newly create slackbuild
This slackbuild was inspired on Archlinux PKGBUILD[1] and Gentoo PF4Public
ebuild[2].

It builds chromium using ungoogled-chromium patches and attempts to use
system libraries as much as possible. Since this is the first release it
might be missing fixes or settings that others have.

Note 1:
To build this you need the slackbuild nodejs-bin installed.

Note 2:
This slackbuild does *not* require 'gn' to be installed, but it has an
extra compilation step before starting the compilation (it uses the 'gn'
bundled with chromium).

[1]: https://github.com/ungoogled-software/ungoogled-chromium-archlinux
[2]: https://github.com/PF4Public/gentoo-overlay/tree/master/www-client/ungoogled-chromium